trim source code of file before checking for sourcemap
the file-change-cache does a sourcemap check, which works by checking if the last line of the file is a sourcemap comment. Sourcemaps sitll work with a trailing newline, and this check was not accounting for that. Just trimming the source to nip any trailing newlines when doing the check fixes it though. Just as en example use case, psc (pursecript compiler) puts a trailing newline on a file. also, Atom by default if you edit a compiled file for debugging will add a trailing newline. 

Fixes: #84